pyproject.toml
[project]
name = "lobbyregister-ingestor"
version = "0.1.1"
description = "Tools to ingest lobby register statistics into Postgres"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "pyyaml>=6.0",
  "python-dateutil",
  "requests>=2.32.5",
  "psycopg[binary,pool]>=3.2",
]

[project.scripts]
ingest-register = "lobbyregister_ingestor.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/lobbyregister_ingestor"]

[tool.hatch.build]
include = ["src/lobbyregister_ingestor/api-docs-lobbyregister.yaml"]
